Add AuthService Change WichtelUserController to get ids from the currently logged in user using AuthService instead of from path variable Remove post endpoint to create new users as this is now done at login

+++ backend/src/main/java/org/example/backend/controller/WichtelUserController.java
@@ -3,8 +3,10 @@ package org.example.backend.controller;
 import lombok.RequiredArgsConstructor;
 import org.example.backend.model.WichtelUser;
 import org.example.backend.model.WichtelUserDTO;
+import org.example.backend.service.AuthService;
 import org.example.backend.service.WichtelUserService;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.oauth2.client.authentication.OAuth2AuthenticationToken;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -16,10 +18,11 @@ import java.util.NoSuchElementException;
 public class WichtelUserController {
 
     private final WichtelUserService service;
+    private final AuthService authService;
 
-    @GetMapping("/{id}")
-    WichtelUser findById(@PathVariable String id){
-        return service.findById(id);
+    @GetMapping()
+    WichtelUser findById(OAuth2AuthenticationToken authentication){
+        return authService.getUserFromAuthToken(authentication);
     }
 
     @GetMapping
@@ -28,23 +31,27 @@ public class WichtelUserController {
     }
 
     @DeleteMapping("/{id}")
-    void deleteById(@PathVariable String id){
-        service.deleteById(id);
+    void deleteById(@PathVariable String id, OAuth2AuthenticationToken authentication){
+        if(authService.loggedInUserHasId(authentication,id)){
+            service.deleteById(id);
+        }
     }
 
-    @PostMapping
-    WichtelUser createUser(@RequestBody WichtelUserDTO dto){
-        return service.save(dto);
-    }
-
-    @PutMapping("/{id}")
-    WichtelUser update(@RequestBody WichtelUserDTO dto,@PathVariable String id){
-        return service.update(dto,id);
+    @PutMapping()
+    WichtelUser update(@RequestBody WichtelUserDTO dto,OAuth2AuthenticationToken authentication){
+        WichtelUser user = authService.getUserFromAuthToken(authentication);
+        return service.update(dto,user.getId());
     }
 
     @ExceptionHandler(NoSuchElementException.class)
     @ResponseStatus(HttpStatus.NOT_FOUND)
     public String handleNotFoundException(NoSuchElementException exception) {
+        return exception.getMessage();
+    }
+
+    @ExceptionHandler(IllegalCallerException.class)
+    @ResponseStatus(HttpStatus.UNAUTHORIZED)
+    public String handleIllegalCallerException(IllegalCallerException exception) {
         return exception.getMessage();
     }
 }
